+++ webview-ui/src/components/chat/AutoApproveMenu.tsx
@@ -1,10 +1,10 @@
 import { VSCodeCheckbox, VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
 import { useCallback, useState } from "react"
 import styled from "styled-components"
-import { useExtensionState } from "../../context/ExtensionStateContext"
-import { AutoApprovalSettings } from "../../../../src/shared/AutoApprovalSettings"
-import { vscode } from "../../utils/vscode"
-import { getAsVar, VSC_FOREGROUND, VSC_TITLEBAR_INACTIVE_FOREGROUND, VSC_DESCRIPTION_FOREGROUND } from "../../utils/vscStyles"
+import { useExtensionState } from "@/context/ExtensionStateContext"
+import { AutoApprovalSettings } from "@shared/AutoApprovalSettings"
+import { vscode } from "@/utils/vscode"
+import { getAsVar, VSC_FOREGROUND, VSC_TITLEBAR_INACTIVE_FOREGROUND, VSC_DESCRIPTION_FOREGROUND } from "@/utils/vscStyles"
 
 interface AutoApproveMenuProps {
 	style?: React.CSSProperties